% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{boxplot_fits_age}
\alias{boxplot_fits_age}
\title{Plots mean ages to all items in dat_name}
\usage{
boxplot_fits_age(dat, dat_name, ylab, mcmcdat = NULL)
}
\arguments{
\item{dat}{description}

\item{dat_name}{description}

\item{ylab}{description}

\item{mcmcdat}{description}
}
\description{
Needs
}
\details{
This function needs a description.
}
