% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{calc_SSB}
\alias{calc_SSB}
\title{Calculate SSB}
\usage{
calc_SSB(type, age_data, VB_params, WL_params, M_params)
}
\arguments{
\item{type}{Type of object. One of c("res", "output_log").}

\item{age_data}{Object containting age frequency data.}

\item{VB_params}{= vector containing parameters for length at age relationship, Von-Bertalanffy.}

\item{WL_params}{= list containing parameters for weight at length relationship.}

\item{M_params}{= list containing parameters for the maturity at age relationship.}
}
\value{
Returns a time series of SSB
}
\description{
Manually calculate SSB from age frequency data.
}
\details{
This function calculates SSB from numbers at age, growth, weight at age, maturity at age.
The input data can be from the OM res object or from the casal output log.
}
