% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{calc_VBweight2}
\alias{calc_VBweight2}
\title{Von Bertalanffy weight 2. This one doesn't require "object" like calc_VBweith()}
\usage{
calc_VBweight2(age1, params1, params2)
}
\arguments{
\item{age1}{age}

\item{params1}{von Bert parameters}

\item{params2}{weight length parameters}
}
\description{
Growth may vary seasonally, therefore the whole (FLQuant-like)
object is passed on
}
\details{
NOTE: params1 and params2 should reference names not position
}
