% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{CASAL_box}
\alias{CASAL_box}
\title{CASAK boxplot}
\usage{
CASAL_box(Plot_mat, YearRange, Ylimits, ablines_h, XYlabels,
  current_y = "")
}
\description{
Needs
}
\details{
This function needs a description.
}
