% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{check_lhps}
\alias{check_lhps}
\title{Checks the life history parameters in the operating model}
\usage{
check_lhps(para)
}
\arguments{
\item{para}{list of operating model parameters}
}
\description{
Compare the specifications of the operating model and assessment model
}
