% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_match}
\alias{check_match}
\title{Check Match}
\usage{
check_match(para)
}
\arguments{
\item{para}{Parameter settings at the point just before running models.}
}
\value{
Returns printed output listing matched/mismatched parameters
}
\description{
Check perfect knowledge between OM and AM
}
\details{
check_match checks if certain parameters match between Planetfish2 and CASAL.
Perfect knowledge between the two models of parameters like life history values, fishing selectivity,
and movement are necessary for a baseline comparison.
}
\examples{
# With default settings:
para = get_om_data()
para = get_casal_para(para)
check_match(para)
}
