% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{create_dataframe}
\alias{create_dataframe}
\title{create data frame}
\usage{
create_dataframe(pred, obs, cv)
}
\arguments{
\item{pred}{description}

\item{obs}{description}

\item{cv}{description}
}
\description{
Needs
}
\details{
This function needs a description.
}
