% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{create_dataframe_tags}
\alias{create_dataframe_tags}
\title{create dataframe tags}
\usage{
create_dataframe_tags(dat_y, yyear)
}
\arguments{
\item{dat}{description}

\item{dat_name}{description}

\item{xlab}{description}

\item{ylab}{description}
}
\description{
Needs
}
\details{
This function needs a description.
}
