% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{cryptic_Biomass}
\alias{cryptic_Biomass}
\title{Cryptic Biomass}
\usage{
cryptic_Biomass(dat = res_quant, dat_name, ages, WatA, mat, firsty,
  plotdims = c(10, 6))
}
\arguments{
\item{dat}{description}

\item{dat_name}{description}

\item{ages}{description}

\item{WatA}{description}

\item{mat}{description}

\item{plotdims}{description}

\item{firstly}{description}
}
\description{
Needs
}
\details{
This function needs a description.
}
