% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{output_SSB}
\alias{output_SSB}
\title{Output SSB}
\usage{
output_SSB(baseline_para, n_tags, n_years_tags, iterations, species,
  casal_path, save_output_path)
}
\arguments{
\item{baseline_para}{Load baseline parameterizaion para into environment.}

\item{n_tags}{Number of tags per year.}

\item{n_years_tags}{Number of years to apply tags.}

\item{iterations}{Number of iterations.}

\item{species}{Species of interest.}

\item{casal_path}{Point to folder with CASAL items.}

\item{save_output_path}{Point to the folder in which to save output.}
}
\value{
Returns a matrix of SSB values generated from the OM and estimated from the AM.
}
\description{
Run operating and assessment models to acquire SSB output.
}
\details{
output_SSB runs the operating model and CASAL assessment model to generate data
for comparison. First and foremost, this is a trial function to streamline the process
of running the models and saving SSB output for Brett's PhD projects. Alterations will be
required if investigations outside the scope of Brett's PhD are desired. This function takes
as input the `para` parameter settings from a particular species. These settings should be the
"default" settings from which adjustment scenarios can be allocated. For example, a typical use could
use Patagonia Toothfish life history parameters, fishing, and sampling specifications reflecting
the stock on  the Kerguelen Plateau. Modifications to the baseline para object related to tagging
or otherwise can be made to facilitate inspection of the impact of hypothetical sampling scenarios.
}
