% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{plot_diagnostics_year_by_age}
\alias{plot_diagnostics_year_by_age}
\title{Residual plots (code by Timothy Earle 2015-08)}
\usage{
plot_diagnostics_year_by_age(dat, dat_name, trans = rep(identity,
  length(dat_name)), new_device = FALSE)
}
\description{
Each year by age
}
\details{
This function needs a description.
}
