% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{plot_fits_in_trellis_tags}
\alias{plot_fits_in_trellis_tags}
\title{Tag fits - not used}
\usage{
plot_fits_in_trellis_tags(dat, dat_name, xlab, ylab)
}
\description{
Needs
}
\details{
This function needs a description.
}
