% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{plot_fits_survey_biomass}
\alias{plot_fits_survey_biomass}
\title{Plot fits survey biomass}
\usage{
plot_fits_survey_biomass(dat, dat_name, xlab, ylab, mcmcdat = NULL)
}
\arguments{
\item{dat}{description}

\item{dat_name}{description}

\item{xlab}{description}

\item{ylab}{description}

\item{mcmcdat}{description}
}
\description{
Needs
}
\details{
This function needs a description.
}
