% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{plot_fits_tagN}
\alias{plot_fits_tagN}
\title{Plot tag-recapture Numbers (by release and recapture year)}
\usage{
plot_fits_tagN(dat, plotdims = c(8, 8))
}
\arguments{
\item{dat}{description}

\item{plotdims}{description}
}
\description{
Needs
}
\details{
This function needs a description.
}
