% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{plot_fits_tagN_area}
\alias{plot_fits_tagN_area}
\title{Plot tag-recapture Numbers (by release and recapture year)}
\usage{
plot_fits_tagN_area(dat, plotdims = c(8, 8), casal_path, runN)
}
\description{
Multi-area:
}
\details{
This function needs a description.
}
