% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{plot_fits_tags_area}
\alias{plot_fits_tags_area}
\title{Plot tag-recapture fits (by release year, recapture year and length bin)}
\usage{
plot_fits_tags_area(dat, plotdims = c(12, 12), casal_path, runN)
}
\arguments{
\item{dat}{description}

\item{plotdims}{description}

\item{casal_path}{description}

\item{runN}{description}
}
\description{
Multi-area:
}
\details{
This function needs a description.
}
