% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{plot_fits_tags_Exp_Obs}
\alias{plot_fits_tags_Exp_Obs}
\title{Plot expected - Observed tag-recapture Numbers in one graph by recapture year}
\usage{
plot_fits_tags_Exp_Obs(dat, plotdims = c(9, 6))
}
\description{
Needs
}
\details{
This function needs a description.
}
