% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{plot_fits_tags_Exp_Obs_area}
\alias{plot_fits_tags_Exp_Obs_area}
\title{Plot expected - Observed tag-recapture Numbers in one graph by recapture year}
\usage{
plot_fits_tags_Exp_Obs_area(dat, plotdims = c(9, 6), casal_path, runN)
}
\arguments{
\item{dat}{description}

\item{plotdims}{description}

\item{casal_path}{description}

\item{runN}{description}
}
\description{
Multi-area:
}
\details{
This function needs a description.
}
