% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{plot_fits_tags_MedianL}
\alias{plot_fits_tags_MedianL}
\title{Plot median length of tag-recapture fits (by release year, recapture year and length bin)}
\usage{
plot_fits_tags_MedianL(dat, plotdims = c(12, 12))
}
\arguments{
\item{dat}{description}

\item{plotdims}{description}
}
\description{
Needs
}
\details{
This function needs a description.
}
