% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{plot_fits_tags_MedianL_area}
\alias{plot_fits_tags_MedianL_area}
\title{Plot median length of tag-recapture fits (by release year, recapture year and length bin)}
\usage{
plot_fits_tags_MedianL_area(dat, plotdims = c(12, 12), casal_path, runN)
}
\description{
Multi-area:
}
\details{
This function needs a description.
}
