% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot residuals from model age/length data fits}
\usage{
plot_residuals(dat, dat_name, ylab, plotdims = c(8, 8))
}
\description{
Needs
}
\details{
This function needs a description.
}
