% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{SSB_err}
\alias{SSB_err}
\title{Relative Error SSB}
\usage{
SSB_err(output, type)
}
\arguments{
\item{output}{The output matrix from a model run.}

\item{type}{Switch. Type of SSB index. Options: "initial", "current", "status".}
}
\value{
Returns a vector distribution of relative error.
}
\description{
Calculate relative error for SSB
}
\details{
This function needs a description.
}
