% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{stock_rec}
\alias{stock_rec}
\title{Stock-recruitment relationship}
\usage{
stock_rec(dat)
}
\arguments{
\item{dat}{description}
}
\description{
Needs
}
\details{
This function needs a description.
}
