% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{calc_SSB_by_region}
\alias{calc_SSB_by_region}
\title{calc SSB by region}
\usage{
calc_SSB_by_region(casal_path, pop_csl, res_quant)
}
\description{
Needs
}
\details{
This function needs a description.
}
