% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{CASAL_box_old}
\alias{CASAL_box_old}
\title{CASAL boxplot old}
\usage{
CASAL_box_old(Plot_mat, YearRange, Ylimits, ablines_h, XYlabels,
  current_y = "")
}
\arguments{
\item{Plot_mat}{description}

\item{YearRange}{description}

\item{Ylimits}{description}

\item{ablines_h}{description}

\item{XYlabels}{description}

\item{current_y}{description}
}
\description{
Needs
}
\details{
This function needs a description.
}
