% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{plot_fits_survey}
\alias{plot_fits_survey}
\title{Plot fits survey}
\usage{
plot_fits_survey(dat, dat_name, xlab, ylab, mcmcdat = NULL)
}
\description{
Needs
}
\details{
This function needs a description.
}
