% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{plot_fits_tags_RobScott}
\alias{plot_fits_tags_RobScott}
\title{Code from Rob Scott (Cefas)}
\usage{
plot_fits_tags_RobScott(dat, dat_name, finalyr, xlab, ylab)
}
\description{
Needs
}
\details{
This function needs a description.
}
