% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phill_Functions_Diagnostics.r
\name{plot_YCS_bubbles}
\alias{plot_YCS_bubbles}
\title{Bubble plots of age observations}
\usage{
plot_YCS_bubbles(est_file, casal_path, yline, dat_name)
}
\description{
Needs
}
\details{
This function needs a description.
}
